+++ app/src/main/java/com/omelan/cofi/components/RecipeItem.kt
@@ -18,13 +18,11 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.google.accompanist.insets.ExperimentalAnimatedInsets
 import com.omelan.cofi.model.Recipe
 import com.omelan.cofi.ui.Spacing
 import com.omelan.cofi.ui.card
 import com.omelan.cofi.ui.shapes
 
-@ExperimentalAnimatedInsets
 @Composable
 fun RecipeItem(recipe: Recipe, onPress: (recipeId: Int) -> Unit) {
     Surface(
@@ -77,7 +75,6 @@ fun RecipeItem(recipe: Recipe, onPress: (recipeId: Int) -> Unit) {
     }
 }
 
-@ExperimentalAnimatedInsets
 @Preview
 @Composable
 fun PreviewRecipeItem() {